requirements.txt
google-genai>=1.0.0
huggingface_hub>=0.24.0
